fix: add missing react and next js packages, move navbar to components
