+++ packages/landing/components/index.ts
@@ -1,0 +1,3 @@
+import { Navbar } from "./navbar/Navbar";
+
+export { Navbar };
