+++ packages/landing/constants/index.ts
@@ -1,1 +1,4 @@
-export const STACKLY_APP_URL = "https://stackly-ui.pages.dev/"; // "https://stackly.eth.limo/";
+export const STACKLY_LANDING_URL =
+  process.env.STACKLY_LANDING_URL ?? "https://stackly.app/";
+export const STACKLY_APP_URL =
+  process.env.STACKLY_APP_URL ?? "https://stackly-ui.pages.dev/"; // "https://stackly.eth.limo/";
