+++ packages/landing/next.config.js
@@ -1,12 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: "export",
-  distDir: "../../dist",
-  reactStrictMode: true,
-  trailingSlash: true,
   webpack(config) {
     config.resolve.fallback = { "pino-pretty": false, lokijs: false };
-
     config.module.rules.push({
       test: /\.svg$/i,
       issuer: /\.[jt]sx?$/,
